fix: prevent unauthorized users from accessing cms ui
Note: This bug fixes a case sensitive issue with unauthorized users
being able to see the CMS UI by accessing /CMS (all caps). The user
would not have been able to see any of the actual content which is
guarded by firestore's security rules.

With this fix, there are additional checks to ensure the user is logged
in and verified to have access (via the site's ACL) before being
rendered the CMS UI.